Cadastro completo de orçamento cotação

+++ Global/Infra/InfraBanco/Modelos/TorcamentoCotacaoOpcaoItemAtomicoCustoFin.cs
@@ -12,12 +12,12 @@ namespace InfraBanco.Modelos
         public int Id { get; set; }
         public int IdItemAtomico { get; set; }
         public int IdOpcaoPagto { get; set; }
-        public decimal DescDado { get; set; }
+        public float DescDado { get; set; }
         public decimal PrecoLista { get; set; }
         public decimal PrecoVenda { get; set; }
         public decimal PrecoNF { get; set; }
-        public decimal CustoFinancFornecCoeficiente { get; set; }
-        public int CustoFinancFornecPrecoListaBase { get; set; }
+        public float CustoFinancFornecCoeficiente { get; set; }
+        public decimal CustoFinancFornecPrecoListaBase { get; set; }
 
         public TorcamentoCotacaoOpcaoItemAtomico TorcamentoCotacaoOpcaoItemAtomico { get; set; }
         public TorcamentoCotacaoOpcaoPagto TorcamentoCotacaoOpcaoPagto { get; set; }
